requirements-test.txt
# Test Dependencies for The Progress Method Bot
# Install with: pip install -r requirements-test.txt

# Core Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.3.0  # For parallel test execution
pytest-html>=3.2.0   # HTML test reports
pytest-json-report>=1.5.0  # JSON test reports

# Testing Utilities
factory-boy>=3.3.0   # Test data factories
faker>=19.3.0        # Fake data generation
freezegun>=1.2.0     # Time mocking
responses>=0.23.0    # HTTP request mocking
httpx>=0.24.0        # Async HTTP client for testing
aioresponses>=0.7.0  # Async HTTP mocking

# Web Testing
flask-testing>=0.8.1
selenium>=4.11.0
webdriver-manager>=3.8.0
beautifulsoup4>=4.12.0

# Database Testing
pytest-postgresql>=5.0.0  # PostgreSQL testing
sqlalchemy-utils>=0.41.0   # Database utilities

# Mocking and Fixtures
unittest-mock>=1.0.0
pytest-factoryboy>=2.5.0
pytest-lazy-fixture>=0.6.0

# Coverage and Reporting
coverage[toml]>=7.2.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0  # Performance benchmarking

# Code Quality (for test linting)
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.4.0

# Async Testing Utilities
pytest-asyncio>=0.21.0
asynctest>=0.13.0; python_version<"3.8"
async-timeout>=4.0.0

# API Testing
requests>=2.31.0
httpx>=0.24.0
pytest-httpx>=0.21.0

# Documentation Testing
doctest>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# Telegram Bot Testing
aiogram>=3.0.0  # Same as main requirements
aiohttp>=3.8.0

# Database Testing
supabase>=2.0.0  # Same as main requirements
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# AI/ML Testing
openai>=1.0.0  # Same as main requirements

# Flask Testing Dependencies
flask>=2.3.0
werkzeug>=2.3.0

# Additional Testing Utilities
tqdm>=4.65.0     # Progress bars for long tests
tabulate>=0.9.0  # Table formatting
colorama>=0.4.6  # Colored terminal output
click>=8.1.0     # CLI testing utilities

# Memory and Performance Testing
memory-profiler>=0.60.0
psutil>=5.9.0

# Fixtures and Sample Data
lorem>=0.1.1     # Lorem ipsum generator
pillow>=10.0.0   # Image processing for test fixtures

# Development and Debugging
pdb++>=0.10.0    # Enhanced debugger
ipython>=8.14.0  # Interactive Python shell
rich>=13.4.0     # Rich text and beautiful formatting

# Test Configuration
toml>=0.10.2     # TOML configuration files
pyyaml>=6.0      # YAML configuration

# Parallel Testing
pytest-xdist[psutil]>=3.3.0

# Type Checking for Tests
types-requests>=2.31.0
types-urllib3>=1.26.0
types-toml>=0.10.0

# Test Reporting
allure-pytest>=2.13.0  # Allure test reports
pytest-clarity>=1.0.0  # Better test failure output

# Security Testing
bandit>=1.7.0    # Security linting
safety>=2.3.0    # Dependency vulnerability checking

# Load Testing (for integration tests)
locust>=2.15.0   # Load testing framework

# Mock Services
responses>=0.23.0   # HTTP mocking
fakeredis>=2.16.0   # Redis mocking
moto>=4.2.0         # AWS services mocking

# Clean up and Test Isolation
pytest-cleandir>=0.1.0
pytest-tempdir>=2019.10.12

# Retry Logic for Flaky Tests
pytest-rerunfailures>=11.1.0

# Test Data Validation
jsonschema>=4.18.0
cerberus>=1.3.0

# Browser Testing (for web interface tests)
playwright>=1.36.0  # Modern browser automation
selenium-wire>=5.1.0  # Extended Selenium

# Database Fixtures
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0